APEXCORE-580 APEXCORE-581 Renamed UserDefinedControlTuple to ControlTuple, moved to different package
