+++ api/src/main/java/org/apache/apex/api/ControlAwareDefaultInputPort.java
@@ -18,6 +18,7 @@
  */
 package org.apache.apex.api;
 
+import org.apache.apex.api.operator.ControlTuple;
 import org.apache.hadoop.classification.InterfaceStability;
 
 import com.datatorrent.api.ControlTupleEnabledSink;
@@ -31,7 +32,7 @@ import com.datatorrent.api.DefaultInputPort;
 public abstract class ControlAwareDefaultInputPort<T> extends DefaultInputPort<T> implements ControlTupleEnabledSink<T>
 {
   @Override
-  public boolean putControl(UserDefinedControlTuple payload)
+  public boolean putControl(ControlTuple payload)
   {
     count++;
     return processControl(payload);
@@ -42,5 +43,5 @@ public abstract class ControlAwareDefaultInputPort<T> extends DefaultInputPort<T
    *
    * @param payload the control tuple payload generated by upstream operator(s)
    */
-  public abstract boolean processControl(UserDefinedControlTuple payload);
+  public abstract boolean processControl(ControlTuple payload);
 }
