+++ api/src/main/java/org/apache/apex/api/operator/ControlTuple.java
@@ -16,15 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.apex.api;
+package org.apache.apex.api.operator;
 
 import org.apache.hadoop.classification.InterfaceStability;
 
 /**
- * Any user generated control tuple must implement {@link UserDefinedControlTuple} interface
+ * Any user generated control tuple must implement {@link ControlTuple} interface
  */
 @InterfaceStability.Evolving
-public interface UserDefinedControlTuple
+public interface ControlTuple
 {
   /**
    * A user generated control tuple must specify a @{@link DeliveryType}
