+++ engine/src/main/java/com/datatorrent/stram/stream/MuxStream.java
@@ -24,7 +24,7 @@ import java.util.HashMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.apex.api.UserDefinedControlTuple;
+import org.apache.apex.api.operator.ControlTuple;
 
 import com.datatorrent.api.Sink;
 import com.datatorrent.stram.engine.Stream;
@@ -123,7 +123,7 @@ public class MuxStream implements Stream.MultiSinkCapableStream
   }
 
   @Override
-  public boolean putControl(UserDefinedControlTuple payload)
+  public boolean putControl(ControlTuple payload)
   {
     put(new CustomControlTuple(payload));
     return false;
